Refactor `simulate_and_update_compute_unit_limit`
